공통컴포넌트 생성
